+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # WaterDrop changelog
 
-## 2.3.4 (Unreleased)
+## 2.4.0 (Unreleased)
 - Small refactor of the DataDog/Statsd listener to align for future extraction to `karafka-common`.
-- Replace `dry-monitor` with home-brew notification layer (API compatible).
+- Replace `dry-monitor` with home-brew notification layer (API compatible) and allow for usage with `ActiveSupport::Notifications`.
 
 ## 2.3.3 (2022-07-18)
 - Replace `dry-validation` with home-brew validation layer and drop direct dependency on `dry-validation`.
