+++ lib/waterdrop/instrumentation/notifications.rb
@@ -1,0 +1,38 @@
+# frozen_string_literal: true
+
+module WaterDrop
+  module Instrumentation
+    # Instrumented is used to hookup external monitoring services to monitor how WaterDrop works
+    class Notifications < WaterDrop::Monitoring::Notifications
+      # List of events that we support in the system and to which a monitor client can hook up
+      # @note The non-error once support timestamp benchmarking
+      EVENTS = %w[
+        producer.closed
+
+        message.produced_async
+        message.produced_sync
+        message.acknowledged
+        message.buffered
+
+        messages.produced_async
+        messages.produced_sync
+        messages.buffered
+
+        buffer.flushed_async
+        buffer.flushed_sync
+
+        statistics.emitted
+
+        error.occurred
+      ].freeze
+
+      private_constant :EVENTS
+
+      # @return [WaterDrop::Instrumentation::Monitor] monitor instance for system instrumentation
+      def initialize
+        super
+        EVENTS.each { |event| register_event(event) }
+      end
+    end
+  end
+end
