+++ spec/lib/waterdrop/monitoring/event_spec.rb
