+++ spec/lib/waterdrop/monitoring/notifications_spec.rb
